[m3nsch] script to more easily run on k8s
Previously users had to `exec` into a pod and manually provide m3nsch
endpoints. This is unnecessary as we can gather the IPs of the agent
pods and pass that via the command line.

Also included a make target to generate the simple manifest bundle in CI
to avoid it becoming stale.